python: raise proper exceptions when an error condition happens
Instead of writing a message to the syslog-ng internal log, raise exceptions
in error cases so that the Python code is able to cope with it.

At the same time syslog-ng internal logs of these events were removed,
exceptions should be noticable enough.  Also, there's a subsequent patch
that exposes Python backtraces in debug mode, which again means that the
internal logs are just duplicating the information.

Even in production mode, the exceptions are reported, only in a condensed
form.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>