python: remove side effect from msg_XXX() calls
Earlier, the Python exception state was cleared with a call to
_py_fetch_and_format_exception(), meaning that the error log was clearing
the exception state, which is not ideal.

This patch changes that, separates the formatting and the clearing of the
exception state, the first becoming a function without side effects, the
second doing nothing but clearing of the exception state.

This is also a basis of a subsequent patch that will give more elaborate
error reports on Python exceptions.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>