driver: make the inner-plugin framework more reusable
We have had the inner-source and inner-dest contexts for implementing
plugins that extend source and destination drivers respectively.

This has worked so that an inner-dest plugin would be invoked during the
initialization of the driver, making it possible for it to change various
LogPipe methods.  The disk buffer plugin was using this mechanism to
override the acquire_queue() method.

However, the way these overrides would get their own state structures was
not solved very well.  I basically added "_data" pointers for all methods
where I anticipated it made sense to override them.  This also added
user_data arguments to acquire_queue/release_queue/queue methods.

This patch gets rid off the various "_data" pointers and the extra argument
for passing those, in favour of using an explicit lookup on the list of
registered plugins.  This lookup now happens via a simple string and the
lookup key is hard-coded into methods that need their own state.  Although
this is a bit less generic, it makes it much easier to override any current
and future methods and makes the code much nicer.

The way we look up instance variables also means, that only one plugin of
the same name can be registered to a specific driver, but I think it is a
safe assumption.

One case that will not be handled and was before is that should two
separate queueing modules register for the same destination, the last
one to be registered would prevail and the ones before would silently
by ignored. Previously this case was explicitly validated and syslog-ng
exited with an error. But this is a small price to pay to make the code
much nicer and the framework capable of doing something other than
queueing mechanisms.

Signed-off-by: Balazs Scheidler <balazs.scheidler@balabit.com>